Oil and Laser Cards
Oil lies around the field for a long while in different shapes depending
on the range, and deals low damage. The laser deals damage all the way
down the row the user is standing in. Lighting was also change to use
List instead of an array to hold the nodes it will hit.